Move ftp related models
