+++ src/main/java/uk/gov/hmcts/reform/sendletter/services/FtpClient.java
@@ -13,9 +13,9 @@ import uk.gov.hmcts.reform.sendletter.exception.FtpException;
 import uk.gov.hmcts.reform.sendletter.logging.AppDependency;
 import uk.gov.hmcts.reform.sendletter.logging.AppDependencyCommand;
 import uk.gov.hmcts.reform.sendletter.logging.Dependency;
+import uk.gov.hmcts.reform.sendletter.model.InMemoryDownloadedFile;
+import uk.gov.hmcts.reform.sendletter.model.Report;
 import uk.gov.hmcts.reform.slc.config.FtpConfigProperties;
-import uk.gov.hmcts.reform.slc.services.steps.sftpupload.InMemoryDownloadedFile;
-import uk.gov.hmcts.reform.slc.services.steps.sftpupload.Report;
 
 import java.io.IOException;
 import java.util.List;
