+++ src/main/java/uk/gov/hmcts/reform/sendletter/services/ReportParser.java
@@ -8,9 +8,9 @@ import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
 import uk.gov.hmcts.reform.sendletter.exception.ReportParsingException;
 import uk.gov.hmcts.reform.sendletter.model.LetterPrintStatus;
+import uk.gov.hmcts.reform.sendletter.model.ParsedReport;
+import uk.gov.hmcts.reform.sendletter.model.Report;
 import uk.gov.hmcts.reform.slc.services.steps.getpdf.FileNameHelper;
-import uk.gov.hmcts.reform.slc.services.steps.sftpupload.ParsedReport;
-import uk.gov.hmcts.reform.slc.services.steps.sftpupload.Report;
 
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
