+++ src/test/java/uk/gov/hmcts/reform/sendletter/services/ReportParserTest.java
@@ -3,8 +3,8 @@ package uk.gov.hmcts.reform.sendletter.services;
 import org.junit.Test;
 import uk.gov.hmcts.reform.sendletter.exception.ReportParsingException;
 import uk.gov.hmcts.reform.sendletter.model.LetterPrintStatus;
-import uk.gov.hmcts.reform.slc.services.steps.sftpupload.ParsedReport;
-import uk.gov.hmcts.reform.slc.services.steps.sftpupload.Report;
+import uk.gov.hmcts.reform.sendletter.model.ParsedReport;
+import uk.gov.hmcts.reform.sendletter.model.Report;
 
 import java.time.ZoneId;
 import java.time.ZonedDateTime;
